put rules and country repo into one
